moved admin and chats to /screen
