+++ src/screens/admin/Ticket/ViewTicket.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import axios from "axios";
-import { useAuth } from "../../context/AuthContext";
-import { BASE_URL } from "../../utils/endPointNames";
+import { useAuth } from "../../../context/AuthContext.jsx";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 import toast from "react-hot-toast";
 const ViewTicket = () => {
   const [auth] = useAuth();
