+++ src/screens/admin/emailTemplates/EmailTemplates.js
@@ -1,13 +1,10 @@
 import React, { useEffect, useState } from "react";
-import { useAuth } from "../../context/AuthContext.jsx";
+import { useAuth } from "../../../context/AuthContext.jsx";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import toast from "react-hot-toast";
-import { BASE_URL } from "../../utils/endPointNames.js";
-import {
-  NEW_EMAILTEMPLATE,
-  VIEW_EMAILTEMPLATE,
-} from "../../utils/routeNames.js";
+import { BASE_URL } from "../../../utils/endPointNames.js";
+import { NEW_EMAILTEMPLATE } from "../../../utils/routeNames.js";
 
 const EmailTemplates = () => {
   const [loader, setLoader] = useState(true);
