+++ src/screens/admin/emailTemplates/NewEmailTemplate.js
@@ -1,6 +1,6 @@
 import React, { useState, useRef } from "react";
 import axios from "axios";
-import { useAuth } from "../../context/AuthContext";
+import { useAuth } from "../../../context/AuthContext.jsx";
 import { useNavigate } from "react-router-dom";
 import JoditEditor from "jodit-react";
 import toast from "react-hot-toast";
@@ -18,7 +18,7 @@ import {
   Col,
   Spinner,
 } from "reactstrap";
-import { BASE_URL } from "../../utils/endPointNames";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 
 function NewEmailTemplate() {
   const [title, setTitle] = useState("");
