+++ src/screens/admin/emailTemplates/ViewEmailTemplate.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect, useRef } from "react";
-import { useAuth } from "../../context/AuthContext";
+import { useAuth } from "../../../context/AuthContext.jsx";
 import { useParams, useNavigate } from "react-router-dom";
 import axios from "axios";
 import JoditEditor from "jodit-react";
-import { BASE_URL } from "../../utils/endPointNames";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 import { toast } from "react-hot-toast";
 
 function ViewEmailTemplate() {
