+++ src/screens/admin/proposal/ProposalDetails.js
@@ -26,8 +26,8 @@ import {
   FaSortAmountDown,
   FaDollarSign,
 } from "react-icons/fa";
-import { BASE_URL } from "../../utils/endPointNames";
-import { useAuth } from "../../context/AuthContext";
+import { BASE_URL } from "../../../utils/endPointNames.js";
+import { useAuth } from "../../../context/AuthContext.jsx";
 import { useParams, useNavigate } from "react-router-dom";
 import toast from "react-hot-toast";
