+++ src/screens/admin/proposal/Proposals.js
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
-import { useAuth } from "../../context/AuthContext";
+import { useAuth } from "../../../context/AuthContext.jsx";
 // import { useEditUserContext } from "../../context/EditUserContext";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
-import * as Routes from "../../utils/routeNames";
+import * as Routes from "../../../utils/routeNames.js";
 
-import { BASE_URL } from "../../utils/endPointNames.js";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 
 function Proposals() {
   const [auth] = useAuth();
