+++ src/screens/admin/subscription/Subscriptions.js
@@ -1,10 +1,10 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
-import { useAuth } from "../../context/AuthContext.jsx";
+import { useAuth } from "../../../context/AuthContext.jsx";
 import { useNavigate } from "react-router-dom";
 import moment from "moment";
-import { BASE_URL } from "../../utils/endPointNames.js";
-import { NEW_SUBCRIPTIONS } from "../../utils/routeNames.js";
+import { BASE_URL } from "../../../utils/endPointNames.js";
+import { NEW_SUBCRIPTIONS } from "../../../utils/routeNames.js";
 
 const Subscriptions = () => {
   const [loader, setLoader] = useState(true);
