+++ src/screens/admin/user/UserDetails.js
@@ -1,11 +1,12 @@
-import axios from "axios";
+
 import React, { useEffect, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
-import { useAuth } from "../../context/AuthContext";
-import { BASE_URL } from "../../utils/endPointNames";
+import { useAuth } from "../../../context/AuthContext.jsx";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 import moment from "moment";
 import toast from "react-hot-toast";
-import ReusableDialog from "../../components/DialogComponent";
+import ReusableDialog from "../../../components/DialogComponent";
+import axios from "axios";
 
 import { Modal, ModalHeader, ModalBody } from "reactstrap";
 function View() {
