+++ src/screens/admin/user/Users.js
@@ -1,10 +1,10 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
-import { useAuth } from "../../context/AuthContext.jsx";
+import { useAuth } from "../../../context/AuthContext.jsx";
 import { useNavigate } from "react-router-dom";
 import toast from "react-hot-toast";
-import * as Routes from "../../utils/routeNames";
-import { BASE_URL } from "../../utils/endPointNames.js";
+import * as Routes from "../../../utils/routeNames.js";
+import { BASE_URL } from "../../../utils/endPointNames.js";
 
 const Users = () => {
   const [auth] = useAuth();
